+++ cli/cli-menu-structures/show-all-previous-active-fasts.js
@@ -1,5 +1,7 @@
 const MenuElement = require('../cli-menu-element');
 const DataServiceSingleton = require('../../data/data-service');
+const {formatDatetimeString} = require('../datetime-helper');
+
 
 // Get singleton instance
 const dataService = new DataServiceSingleton().instance;
@@ -15,19 +17,14 @@ module.exports = class AllPreviousFastsMenu {
             for (let i = 0; i < allPreviousFasts.length; i++) {
                 this._menu[( i * 6)] = new MenuElement('', '--------------------------');
                 this._menu[1+ ( i * 6)] = new MenuElement('Status', allPreviousFasts[i]._status);
-                this._menu[2 + (i * 6)] = new MenuElement('Started', this.formatDatetimeString(allPreviousFasts[i]._started));
-                this._menu[3 + (i * 6)] = new MenuElement('Ended', this.formatDatetimeString(allPreviousFasts[i]._ending));
+                this._menu[2 + (i * 6)] = new MenuElement('Started', formatDatetimeString(allPreviousFasts[i]._started));
+                this._menu[3 + (i * 6)] = new MenuElement('Ended', formatDatetimeString(allPreviousFasts[i]._ending));
                 this._menu[4 + (i * 6)] = new MenuElement('Type', `${allPreviousFasts[i]._type}h`);
                 this._menu[5 + (i * 6)] = new MenuElement('', '--------------------------')
             }
         } else {
             this._menu[0] = new MenuElement('', '\nThere are no previously saved fast sessions.\n');
         }
-    }
-
-    formatDatetimeString = (dateTime) => {
-        const parsedDateTime = new Date(dateTime);
-        return (`${parsedDateTime.toLocaleDateString('en-GB')} ${parsedDateTime.toLocaleTimeString('en-GB')}`);
     }
 
     get menu() {
